fix(runtime): handle pre-parsed Node request bodies
